refactor(#38): remove unused controllers
