--- app/Http/Controllers/AuthorController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class AuthorController extends Controller
-{
-    //
-}
