--- app/Http/Controllers/BlockController.php
@@ -1,49 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\Block;
-use Illuminate\Http\Request;
-
-class BlockController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(Block $block)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, Block $block)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Block $block)
-    {
-        //
-    }
-}
