--- app/Http/Controllers/ChapterController.php
@@ -1,49 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\Chapter;
-use Illuminate\Http\Request;
-
-class ChapterController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(Chapter $chapter)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, Chapter $chapter)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Chapter $chapter)
-    {
-        //
-    }
-}
